Extract tab label component
