+++ src/games/game-settings/gameWithSettings/playerInputsPresets/tabLabel/Props.ts
@@ -1,0 +1,6 @@
+interface Props {
+  name: string
+  onChange: (newName: string) => void
+}
+
+export default Props
